// src/types/checkout.ts
export type CountriesT = {
    name: string
    value: string
}

export type PaymentMethodsT = {
    icon?: string
    image?: string
    name: string
}

export type GroupIconsT = {
    icon?: string
    image?: string
}

export type ContactDetailT = {
    cardNumber?: string
    expiration?:string
}